+++ tests/agent/test_k8s_agent.py
@@ -1,3 +1,5 @@
+from os import path
+import tempfile
 from unittest.mock import MagicMock
 
 import pytest
@@ -7,6 +9,7 @@ pytest.importorskip("kubernetes")
 import yaml
 
 from prefect.agent.kubernetes import KubernetesAgent
+from prefect.agent.kubernetes.agent import check_heartbeat
 from prefect.environments.storage import Docker, Local
 from prefect.utilities.configuration import set_temporary_config
 from prefect.utilities.graphql import GraphQLResult
@@ -213,3 +216,70 @@ def test_k8s_agent_generate_deployment_yaml_no_resource_manager(
     assert agent_env[2]["value"] == "test_namespace"
 
     assert len(deployment["spec"]["template"]["spec"]["containers"]) == 1
+
+
+def test_k8s_agent_heartbeat_creates_file(monkeypatch, runner_token):
+    k8s_config = MagicMock()
+    monkeypatch.setattr("kubernetes.config", k8s_config)
+
+    with tempfile.TemporaryDirectory() as tempdir:
+
+        monkeypatch.setattr(
+            "prefect.agent.kubernetes.agent.AGENT_DIRECTORY",
+            "{}/.prefect/agent".format(tempdir),
+        )
+
+        agent = KubernetesAgent()
+        agent.heartbeat()
+
+        assert path.exists("{}/.prefect/agent/heartbeat".format(tempdir))
+
+
+def test_k8s_agent_heartbeat_modifies(monkeypatch, runner_token):
+    k8s_config = MagicMock()
+    monkeypatch.setattr("kubernetes.config", k8s_config)
+
+    with tempfile.TemporaryDirectory() as tempdir:
+
+        monkeypatch.setattr(
+            "prefect.agent.kubernetes.agent.AGENT_DIRECTORY",
+            "{}/.prefect/agent".format(tempdir),
+        )
+
+        agent = KubernetesAgent()
+        agent.heartbeat()
+
+        assert path.exists("{}/.prefect/agent/heartbeat".format(tempdir))
+
+        first = path.getmtime("{}/.prefect/agent/heartbeat".format(tempdir))
+
+        # Wait one second until next heartbeat
+        import time
+
+        time.sleep(1)
+
+        agent.heartbeat()
+        second = path.getmtime("{}/.prefect/agent/heartbeat".format(tempdir))
+
+        assert second > first
+
+
+def test_k8s_agent_heartbeat_check(monkeypatch, runner_token):
+    k8s_config = MagicMock()
+    monkeypatch.setattr("kubernetes.config", k8s_config)
+
+    with tempfile.TemporaryDirectory() as tempdir:
+
+        monkeypatch.setattr(
+            "prefect.agent.kubernetes.agent.AGENT_DIRECTORY",
+            "{}/.prefect/agent".format(tempdir),
+        )
+
+        agent = KubernetesAgent()
+
+        # Verify no exit codes of status `1` were raised
+        agent.heartbeat()
+        assert not check_heartbeat()
+
+        agent.heartbeat()
+        assert not check_heartbeat()